Первичная попытка задания 5.16
